requirements.txt
flask
flask-sqlalchemy
python-magic
flask-limiter
pymemcache
pytesseract
Pillow
